backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-generativeai>=0.3.2
pydantic>=2.8.0
python-dotenv>=1.0.0
starlette>=0.27.0
requests>=2.31.0